+++ packages/ragbits-core/src/ragbits/core/llms/factory.py
@@ -13,23 +13,12 @@ def get_llm_from_factory(factory_path: str) -> LLM:
         factory_path (str): The path to the factory function.
 
     Returns:
-        LLM: An instance of the LLM.
+        LLM: An instance of the LLM class.
     """
     module_name, function_name = factory_path.rsplit(".", 1)
     module = importlib.import_module(module_name)
     function = getattr(module, function_name)
     return function()
-
-
-def has_default_llm(llm_type: LLMType = LLMType.TEXT) -> bool:
-    """
-    Check if the default LLM factory is set in the configuration.
-
-    Returns:
-        bool: Whether the default LLM factory is set.
-    """
-    default_factory = core_config.default_llm_factories.get(llm_type, None)
-    return default_factory is not None
 
 
 def get_default_llm(llm_type: LLMType = LLMType.TEXT) -> LLM:
@@ -43,15 +32,8 @@ def get_default_llm(llm_type: LLMType = LLMType.TEXT) -> LLM:
     Returns:
         LLM: An instance of the default LLM.
 
-    Raises:
-        ValueError: If the default LLM factory is not set or expected llm type is not defined in config
     """
-    if llm_type not in core_config.default_llm_factories:
-        raise ValueError(f"Default LLM of type {llm_type} is not defined in pyproject.toml config.")
     factory = core_config.default_llm_factories[llm_type]
-    if factory is None:
-        raise ValueError("Default LLM factory is not set")
-
     return get_llm_from_factory(factory)
 
 
@@ -61,7 +43,7 @@ def simple_litellm_factory() -> LLM:
     default options, and assumes that the API key is set in the environment.
 
     Returns:
-        LLM: An instance of the LiteLLM.
+        LLM: An instance of the LiteLLM class.
     """
     return LiteLLM()
 
@@ -72,6 +54,16 @@ def simple_litellm_vision_factory() -> LLM:
     default options, and assumes that the API key is set in the environment.
 
     Returns:
-        LLM: An instance of the LiteLLM.
+        LLM: An instance of the LiteLLM class.
     """
     return LiteLLM(model_name="gpt-4o-mini")
+
+
+def simple_litellm_structured_output_factory() -> LLM:
+    """
+    A basic LLM factory that creates an LiteLLM instance with the support for structured output.
+
+    Returns:
+        LLM: An instance of the LiteLLM class.
+    """
+    return LiteLLM(model_name="gpt-4o-mini-2024-07-18", use_structured_output=True)