+++ packages/ragbits-guardrails/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.4.0 (2024-11-27)
+
+### Changed
+
+- ragbits-core updated to version v0.4.0
+
 ## 0.3.0 (2024-11-06)
 
 ### Added